syz-dash: move patch parsing to pkg/kernel
Patch parsing is not dashboard-specific and can
be reused by other programs.
